Fix compilation warnings across multiple modules
- Ivr.cpp: initialize dict to NULL to fix -Wmaybe-uninitialized
- ModXml.cpp: remove deprecated initGenericErrorDefaultFunc/xmlKeepBlanksDefault,
  pass XML_PARSE_NOBLANKS to xmlReadMemory instead
- raw_sock.cpp: fix SAv4(&from) to SAv4(from) - was incorrectly casting
  address of pointer variable instead of the sockaddr_storage pointer
- AmRtpPacket.cpp: use memcpy instead of type-punning cast to fix
  -Wstrict-aliasing
- udp_trsp.cpp: add void* intermediate cast in CMSG_DATA macros to fix
  -Wstrict-aliasing
- RpcPeer.cpp: compute ns_begin from snd_size instead of negative offset
  from msgbuf to fix -Warray-bounds
- fct.h: cast to int in fct_chk_eq_int macro to fix -Wsign-compare

+++ core/tests/fct.h
@@ -3828,10 +3828,10 @@ _fct_chk_full_str(char const *s)
 
 #define fct_chk_eq_int(V1, V2) \
     fct_xchk(\
-        ((V1) == (V2)),\
+        ((int)(V1) == (int)(V2)),\
         "chq_eq_int: %d != %d",\
-        (V1),\
-        (V2)\
+        (int)(V1),\
+        (int)(V2)\
         )
